gah, missed a save

+++ parts/components/HD44780.cpp
@@ -266,13 +266,13 @@ uint32_t HD44780::ProcessWrite()
 
 	if (four) { // 4 bits !
 		if (comp)
-			m_uiDataPins = (m_uiDataPins & 0xf0U) | ((m_uiPinState >>  (unsigned int)D4) & 0xfU);
+			m_uiDataPins = ((unsigned)m_uiDataPins & 0xf0U) | (((unsigned)m_uiPinState >> (unsigned)D4) & 0xfU);
 		else
-			m_uiDataPins = (m_uiDataPins & 0xfU) | ((m_uiPinState >>  (unsigned int)(D4-4)) & 0xf0U);
+			m_uiDataPins = ((unsigned)m_uiDataPins & 0xfU) | (((unsigned)m_uiPinState >> (unsigned)(D4-4U)) & 0xf0U);
 		write = comp;
 		ToggleFlag(HD44780_FLAG_LOWNIBBLE);
 	} else {	// 8 bits
-		m_uiDataPins = (m_uiPinState >>  (unsigned int)D0) & 0xffU;
+		m_uiDataPins = ((unsigned)m_uiPinState >> (unsigned int)D0) & 0xffU;
 		write++;
 	}
 	RaiseIRQ(DATA_IN, m_uiDataPins);
